SCSS
In navigation.scss file started to make the styles for  the button when
closed and when opened, using in the last one a new class called open
that isnt included in html.
JAVASCRIPT
In javascript added an event of click in the menu-btn which will add the
'open' class if sdoesnt exist the class, and remove the class if already
exists ,to show the behaviour of the menu hamburguer that transforms to
a cross when the open class exists and return to burguer menu when the
class is removed.
SCSS
Added styles to the menu list background color and color of the text
including  when its hovered.
NEXT STEP
is to include visbility or transform YX out of the screen using negative numbers zero of the list at this stage and  next add a open class that
will show the menu list with visibility. Add the functionality to the
javasdcript to permit that this elementos have also a class called open
that when exists will show the menu and when its removed the menu will
be hidden .